requirements.txt
numpy
omegaconf
opencv-python
opencv-python-headless
pillow
timm
torch
torchvision
transformers
accelerate
pymupdf
google-generativeai
langfuse
ipykernel
docling
black
isort
flake8
pytest
